libs/analitiq/requirements_cloud.txt
langchain==0.2.1
langchain_aws==0.1.6
langchain-community==0.2.1
langchain-core==0.2.1
langchain-mistralai==0.1.7
langchain-text-splitters==0.2.0
nltk==3.9.1
pandas==1.5.0
pyyaml==6.0.1
redshift-connector==2.1.1
sqlparse==0.5.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tabulate==0.9.0
weaviate-client==4.5.6
psycopg2==2.9.9